fix(thumbnail): fixed thumbnails on CI
There was a bug in the "route-thumbnail-plugin" that prevented the processed
files from being copied to the dist folder. Made a configurable "locations"-
property to allow for multiple output locations.

Configured the plugin to output in the dist/static/assets folder, as well
as next to the blog-posts themselves. This is for "ng serve" to see them
without having to do static resource processing first.