docs: adding guide on how blog posts should be created
Added details in the main README.md-file on how blog posts should be
created. This includes naming conventions, and image dimensions.

In addition to this, I also added documentation on how the plugins work
(and what they do). This is however just a first initial stab at this, and
not done, by any means.