feat(post): added footer with feedback mechanism
Added a post footer component (the old one got renamed to "statusbar").
The footer now contains buttons for reporting content improvements and bugs.
There's been made room for displaying "references" (in said header), but
that's not been populated yet.

To configure GitHub issues links added a means of configuration in the
environments file (and model).