fix(license): fixed issue with matching correct licenses
Found out I'd been using Fuse.js the "wrong way around". I'd always though
that a higher score was a better match, however the correct behaviour is
that:
- 0 is a perfect match
- 1 is a complete mismatch

Also added support for matching upon "public domain", which means that
there is actually no license. So had to add that as a special case in the
code.